Move Constants to internal package.
